Add RNFR hotfix for libunftp
In https://github.com/bolcom/libunftp/pull/539 I've created PR that adds
behaviour to the RNFR command that is necessary for some FTP clients:
They use this command to check if a file exists, but the library always
returned a result that indicated that it did, even if the file didn't
exist.

I have a use case where my scanner uses FTP to upload scans, and it uses
this trick. If I try to scan a file it responds that the file already
exists, so the FTP server needs a new response code when the file
doesn't exist yet. With this patch applied, it works. Once the change
has been merged, the patch can be removed.